+++ MainGUI/Resources/Changelog.md
@@ -1,10 +1,11 @@
 Modnix Changelog
 
-# Version 2.5.7, in-development
+# Version 2.5.7, 2021-02-19
 
 * New: Support GOG Galaxy online launch. Thanks mad2342 for helping with detection.
 * New: Detect Phoenix Point 1.10 Orryx version.
 * Fix: Fix game path detection error when Steam is not installed for current user.
+* Fix: Make sure Add Mods button is enabled on non-standard installation path.
 
 # Version 2.5.6, 2021-01-25